+++ reaktive/src/commonMain/kotlin/com/badoo/reaktive/base/SubscribeSafe.kt
@@ -1,8 +1,8 @@
-package com.badoo.reaktive.maybe
+package com.badoo.reaktive.base
 
 import com.badoo.reaktive.utils.handleSourceError
 
-internal fun <T> Maybe<T>.subscribeSafe(observer: MaybeObserver<T>) {
+internal fun <T> Source<T>.subscribeSafe(observer: T) where T : Observer, T : ErrorCallback {
     try {
         subscribe(observer)
     } catch (e: Throwable) {
