+++ reaktive/src/commonMain/kotlin/com/badoo/reaktive/completable/AsMaybe.kt
@@ -4,6 +4,7 @@ import com.badoo.reaktive.disposable.Disposable
 import com.badoo.reaktive.disposable.DisposableWrapper
 import com.badoo.reaktive.maybe.Maybe
 import com.badoo.reaktive.maybe.maybeUnsafe
+import com.badoo.reaktive.base.subscribeSafe
 
 fun <T> Completable.asMaybe(): Maybe<T> =
     maybeUnsafe { observer ->
