+++ reaktive/src/commonMain/kotlin/com/badoo/reaktive/completable/AsObservable.kt
@@ -4,6 +4,7 @@ import com.badoo.reaktive.disposable.Disposable
 import com.badoo.reaktive.disposable.DisposableWrapper
 import com.badoo.reaktive.observable.Observable
 import com.badoo.reaktive.observable.observableUnsafe
+import com.badoo.reaktive.base.subscribeSafe
 
 fun <T> Completable.asObservable(): Observable<T> =
     observableUnsafe { observer ->
