+++ reaktive/src/commonMain/kotlin/com/badoo/reaktive/completable/DoOnBefore.kt
@@ -5,6 +5,7 @@ import com.badoo.reaktive.base.ErrorCallback
 import com.badoo.reaktive.disposable.Disposable
 import com.badoo.reaktive.disposable.DisposableWrapper
 import com.badoo.reaktive.disposable.wrap
+import com.badoo.reaktive.base.subscribeSafe
 import com.badoo.reaktive.utils.atomicreference.AtomicReference
 
 fun Completable.doOnBeforeSubscribe(action: (Disposable) -> Unit): Completable =
