+++ reaktive/src/commonMain/kotlin/com/badoo/reaktive/completable/ObserveOn.kt
@@ -2,6 +2,7 @@ package com.badoo.reaktive.completable
 
 import com.badoo.reaktive.disposable.CompositeDisposable
 import com.badoo.reaktive.disposable.Disposable
+import com.badoo.reaktive.base.subscribeSafe
 import com.badoo.reaktive.scheduler.Scheduler
 
 fun Completable.observeOn(scheduler: Scheduler): Completable =
