--- reaktive/src/commonMain/kotlin/com/badoo/reaktive/completable/SubscribeSafe.kt
@@ -1,11 +1,0 @@
-package com.badoo.reaktive.completable
-
-import com.badoo.reaktive.utils.handleSourceError
-
-internal fun Completable.subscribeSafe(observer: CompletableObserver) {
-    try {
-        subscribe(observer)
-    } catch (e: Throwable) {
-        handleSourceError(e, observer::onError)
-    }
-}
