+++ reaktive/src/commonMain/kotlin/com/badoo/reaktive/maybe/AsCompletable.kt
@@ -5,6 +5,7 @@ import com.badoo.reaktive.completable.CompletableCallbacks
 import com.badoo.reaktive.completable.completableUnsafe
 import com.badoo.reaktive.disposable.Disposable
 import com.badoo.reaktive.disposable.DisposableWrapper
+import com.badoo.reaktive.base.subscribeSafe
 
 fun Maybe<*>.asCompletable(): Completable =
     completableUnsafe { observer ->
