+++ reaktive/src/commonMain/kotlin/com/badoo/reaktive/maybe/AsObservable.kt
@@ -5,6 +5,7 @@ import com.badoo.reaktive.disposable.Disposable
 import com.badoo.reaktive.disposable.DisposableWrapper
 import com.badoo.reaktive.observable.Observable
 import com.badoo.reaktive.observable.observableUnsafe
+import com.badoo.reaktive.base.subscribeSafe
 
 fun <T> Maybe<T>.asObservable(): Observable<T> =
     observableUnsafe { observer ->
