+++ reaktive/src/commonMain/kotlin/com/badoo/reaktive/maybe/SubscribeOn.kt
@@ -2,6 +2,7 @@ package com.badoo.reaktive.maybe
 
 import com.badoo.reaktive.disposable.CompositeDisposable
 import com.badoo.reaktive.disposable.Disposable
+import com.badoo.reaktive.base.subscribeSafe
 import com.badoo.reaktive.scheduler.Scheduler
 
 fun <T> Maybe<T>.subscribeOn(scheduler: Scheduler): Maybe<T> =
