+++ reaktive/src/commonMain/kotlin/com/badoo/reaktive/maybe/Susbcribe.kt
@@ -2,6 +2,7 @@ package com.badoo.reaktive.maybe
 
 import com.badoo.reaktive.disposable.Disposable
 import com.badoo.reaktive.disposable.DisposableWrapper
+import com.badoo.reaktive.base.subscribeSafe
 import com.badoo.reaktive.utils.UseReturnValue
 import com.badoo.reaktive.utils.handleSourceError
