+++ reaktive/src/commonMain/kotlin/com/badoo/reaktive/observable/Concat.kt
@@ -1,5 +1,6 @@
 package com.badoo.reaktive.observable
 
+import com.badoo.reaktive.base.subscribeSafe
 import com.badoo.reaktive.disposable.Disposable
 import com.badoo.reaktive.disposable.DisposableWrapper
 import com.badoo.reaktive.utils.atomicreference.AtomicReference
