+++ reaktive/src/commonMain/kotlin/com/badoo/reaktive/single/AsCompletable.kt
@@ -4,6 +4,7 @@ import com.badoo.reaktive.base.ErrorCallback
 import com.badoo.reaktive.base.Observer
 import com.badoo.reaktive.completable.Completable
 import com.badoo.reaktive.completable.completableUnsafe
+import com.badoo.reaktive.base.subscribeSafe
 
 fun Single<*>.asCompletable(): Completable =
     completableUnsafe { observer ->
