+++ reaktive/src/commonMain/kotlin/com/badoo/reaktive/single/AsMaybe.kt
@@ -5,6 +5,7 @@ import com.badoo.reaktive.base.Observer
 import com.badoo.reaktive.base.SuccessCallback
 import com.badoo.reaktive.maybe.Maybe
 import com.badoo.reaktive.maybe.maybeUnsafe
+import com.badoo.reaktive.base.subscribeSafe
 
 fun <T> Single<T>.asMaybe(): Maybe<T> =
     maybeUnsafe { observer ->
