+++ reaktive/src/commonMain/kotlin/com/badoo/reaktive/single/FlatMapMaybe.kt
@@ -6,7 +6,7 @@ import com.badoo.reaktive.disposable.DisposableWrapper
 import com.badoo.reaktive.maybe.Maybe
 import com.badoo.reaktive.maybe.MaybeObserver
 import com.badoo.reaktive.maybe.maybeUnsafe
-import com.badoo.reaktive.maybe.subscribeSafe
+import com.badoo.reaktive.base.subscribeSafe
 
 fun <T, R> Single<T>.flatMapMaybe(mapper: (T) -> Maybe<R>): Maybe<R> =
     maybeUnsafe { observer ->
